Handmade Hero Day2
First FFI code written in Rust. It does not yet expose a safe interface instead uses the ffi directly. Needs refactoring later when the Platform segment is done.
Creates a Window with a Title and implements a very basic MessageQueue which Handles redrawing
the windows Contents with the Color White. All FFI is written by hand!