Fixed hard to find Bug on entering chunk x: 23 y: 11
As soon as the player was standing on entering this chunk there was a
tree which got assigned the exact same slot in the hashmap of our
sim-region. So the player simply "vanished".
To hot-fix this issue i just implemented an additional check to see if
the item found in the hashmap has the same storage_index as the one we
want to find. This is not a perfect solution because the item in the hashmap
will still get overriden. All the code is still evolving though so i left it
as is and added a Todo if the code should stay this way till the end!